Split source/trigger validation into separate files
