+++ ts/src/header-validator/filters-index.ts
@@ -2,7 +2,9 @@ import { SourceType } from '../source-type'
 import { Context } from './context'
 import * as filters from './filters'
 import { makeLi } from './issue-utils'
-import { validateJSON, filterPair, filterData } from './validate-json'
+import { validateJSON } from './validate-json'
+import { filterData } from './validate-source'
+import { filterPair } from './validate-trigger'
 
 const form = document.querySelector<HTMLFormElement>('form')!
 const sourceAgeInput = form.elements.namedItem(
