+++ ts/src/header-validator/validator.ts
@@ -1,29 +1,9 @@
 import { ValidationResult } from './context'
 import { Maybe } from './maybe'
-import { Source } from './source'
-import { Trigger } from './trigger'
-import * as toJson from './to-json'
-import * as json from './validate-json'
 
 export interface Validator<T> {
   validate(input: string): [ValidationResult, Maybe<T>]
   serialize(value: T): string
-}
-
-export function source(opts: Readonly<json.SourceOptions>): Validator<Source> {
-  return {
-    validate: (input) => json.validateSource(input, opts),
-    serialize: (value) => toJson.serializeSource(value, opts),
-  }
-}
-
-export function trigger(
-  opts: Readonly<json.RegistrationOptions>
-): Validator<Trigger> {
-  return {
-    validate: (input) => json.validateTrigger(input, opts),
-    serialize: (value) => toJson.serializeTrigger(value, opts),
-  }
 }
 
 export interface Output extends ValidationResult {
